requirements.txt
streamlit
pandas
matplotlib
google-generativeai
python-dotenv